+++ Editor/ALFBTTextField.cs
@@ -1,9 +1,0 @@
-﻿using System;
-
-namespace Cobilas.Unity.Editor.Management.Translation {
-    [Serializable]
-    public sealed class ALFBTTextField {
-        public string Name;
-        public string Text;
-    }
-}