+++ Runtime/TranslationManager.cs
@@ -8,6 +8,7 @@ using Cobilas.Unity.Management.Resources;
 using Cobilas.Unity.Management.RuntimeInitialize;
 #if UNITY_EDITOR
 using UnityEditor;
+using Cobilas.Unity.Management.Build;
 #endif
 
 namespace Cobilas.Unity.Management.Translation {
@@ -17,20 +18,51 @@ namespace Cobilas.Unity.Management.Translation {
 #if UNITY_EDITOR
         [InitializeOnLoadMethod]
         private static void InitEditor() {
+            CobilasBuildProcessor.EventOnPreprocessBuild += (pp, br) => {
+                if (pp == CobilasEditorProcessor.PriorityProcessor.High)
+                    Refresh();
+            };
+            CobilasEditorProcessor.projectChanged += (pp) => {
+                if (pp == CobilasEditorProcessor.PriorityProcessor.Low)
+                    Refresh();
+            };
             if (!EditorApplication.isPlaying) return;
             Reset();
             Init();
         }
-#endif
+
+        [MenuItem("Tools/Translation Manager/Refresh Translation Manager")]
+        [CRIOLM_CallWhen(typeof(CobilasResources), CRIOLMType.BeforeSceneLoad)]
+        private static void Refresh() {
+            Debug.Log(string.Format("[TranslationManager]Refresh [{0}]", System.DateTime.Now));
+            TranslationList.Refresh();
+        }
+
+        [MenuItem("Tools/Translation Manager/Create Translation Folder")]
+        private static void CreateTranslationFolder() {
+            string path = Path.Combine(Application.dataPath, "Resources/Translation");
+            if (!Directory.Exists(path)) {
+                Directory.CreateDirectory(path);
+                AssetDatabase.Refresh();
+            }
+        }
+
+        [CRIOLM_CallWhen(typeof(CobilasResources), CRIOLMType.AfterSceneLoad)]
+#else
         [CRIOLM_BeforeSceneLoad]
+#endif
         private static void Init() {
             Application.quitting += management.Dispose;
-            TextAsset[] texts = CobilasResources.GetAllSpecificObjectInFolder<TextAsset>("Translation");
-            for (int I = 0; I < ArrayManipulation.ArrayLength(texts); I++) {
-                Debug.Log(texts[I].name);
-                using (ALFBTRead read = ALFBTRead.Create(new StringReader(texts[I].text)))
-                    Load(read);
-            }
+            TranslationList[] list = CobilasResources.GetAllSpecificObjectInFolder<TranslationList>("Resources/Translation");
+            for (int I = 0; I < ArrayManipulation.ArrayLength(list); I++)
+                foreach (var item in list[I])
+                    if (item != null)
+                        using (item.Stream) {
+                            item.Stream.Seek(0, SeekOrigin.Begin);
+                            using (ALFBTRead read = ALFBTRead.Create(item.Stream))
+                                if (Load(read))
+                                    Debug.Log("[MemoryStream]ALFBT load");
+                        }
         }
 
         public static void Reset() {
@@ -38,7 +70,7 @@ namespace Cobilas.Unity.Management.Translation {
             management = new TranslationManagement();
         }
 
-        public static void Load(ALFBTRead read) => management.LoadTranslation(read);
+        public static bool Load(ALFBTRead read) => management.LoadTranslation(read);
 
         public static TextFlag GetTextFlag(string path) => management.GetTextFlag(path);