+++ CHANGELOG.md
@@ -1,6 +1,36 @@
 # Changelog
+## [1.0.3] 13/08/2022
+- Change Cobilas MG Translations.asset
+- Change Runtime\ALFBTLanguage.cs
+- Change Runtime\TranslationManager.cs
+## [1.0.3] 13/08/2022
+### Change
+- Change Editor\Cobilas.Unity.Editor.Management.Translations.asmdef
+- Change Runtime\Cobilas.Unity.Management.Translations.asmdef
+- Change Runtime\TranslationManager.cs
+### Add
+- Add Runtime\TranslationList.cs
+- Add Runtime\ALFBTTextFlag.cs
+- Add Runtime\ALFBTMarkingFlag.cs
+- Add Runtime\ALFBTLanguage.cs
+- Add Runtime\ALFBTFlagBase.cs
+- Add Editor\TranslationListInspector.cs
+- Add Editor\ALFBTTextFlagInspector.cs
+- Add Editor\ALFBTMarkingFlagInspector.cs
+- Add Editor\ALFBTLanguageInspector.cs
+### Remove
+- Remove Editor\ALFBTObject.cs
+- Remove Editor\ALFBTTextField.cs
+## [1.0.3] 09/08/2022
+- Change CHANGELOG.md
+- Change Editor\ALFBTObject.cs
+## [1.0.3] 08/08/2022
+- Fix CHANGELOG.md
+- Fix package.json
+- Change Editor\ALFBTObject.cs
+- Change Runtime\TranslationManager.cs
 ## [1.0.2] 31/07/2022
-- Add CHANGELOG.md
+- Fix CHANGELOG.md
 - Fix package.json
 - Add Cobilas MG Translations.asset
 - Remove Runtime\DependencyWarning.cs